requirements.txt
streamlit
numpy
faiss-cpu
transformers
groq
torch